erp/requirements.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
diff-match-patch==20121119
Django==1.11.7
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.0.0
djangorestframework==3.8.2
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
idna==2.6
itchat==1.3.10
jdcal==1.4
Markdown==2.6.9
numpy==1.14.3
odfpy==1.3.6
openpyxl==2.5.2
pandas==0.23.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
PyMySQL==0.7.11
pypng==0.0.18
PyQRCode==1.2.1
python-dateutil==2.7.3
pytz==2017.3
PyYAML==3.12
requests==2.18.4
reversion==0.2
revision==0.1.1
schedule==0.5.0
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
sklearn==0.0
tablib==0.12.1
toml==0.9.4
unicodecsv==0.14.1
urllib3==1.22
xadmin==0.6.1
xlrd==1.1.0
xlwt==1.3.0
